fix: historical_data.volume overflowed int4, poisoning whole insert batches
Bug report: Postgres logged "value '3200578500' is out of range for
type integer" repeatedly, every 30 minutes, matching this app's own
cron sweep (server.js: cron.schedule('*/30 * * * *', ...)). Verified
the report's mechanism is exactly right:

- historical_data.volume was INTEGER (max ~2.147 billion). Yahoo
  returns volume figures for some futures that exceed that.
- The Yahoo/IBKR/continuous-series batch inserts all use a single
  `INSERT ... SELECT unnest($n::integer[])` statement wrapped in one
  transaction, for every bar fetched in that call. One oversized
  volume value fails the whole statement (Postgres validates the
  array cast before any row is written), so the entire batch's other
  bars silently fail to persist too — not just the offending ones.
  Confirmed the failure is in the explicit `::integer[]` cast itself,
  not just the column type, so widening the column alone wasn't
  enough.

Fix: widened historical_data.volume to BIGINT (schema + migration in
modules/database.js, following this file's existing idempotent-ALTER
migration pattern), and changed all three `unnest($n::integer[])`
volume casts in historical-data-service.js to `::bigint[]`. The
tradingview-webhook.js single-row insert needed no cast change (no
explicit type cast there) but its comment was stale, fixed too.

Applied the schema migration directly against the live DB rather than
waiting for a restart, so the already-failing sweep would recover
immediately — historical_data turned out to be ~59.4M rows, so the
ALTER's required full-table rewrite took about 6 minutes and queued
several concurrent app inserts behind its lock (they drained
successfully once it completed, no data loss). Verified the actual
fix afterward by round-tripping the exact query pattern from the code
with one of the real reported oversized volume values
(3,200,578,500) — it now inserts successfully.

+++ modules/historical-data-service.js
@@ -520,7 +520,7 @@ async function populateHistoricalData(task, broadcastStatus, wss) { // Task obje
 
                     const query = `
                         INSERT INTO historical_data (futures_setting_id, contract_month, time, open, high, low, close, volume, timeframe, source, is_continuous)
-                        SELECT $1, $2, unnest($3::timestamptz[]), unnest($4::numeric[]), unnest($5::numeric[]), unnest($6::numeric[]), unnest($7::numeric[]), unnest($8::integer[]), $9, 'Yahoo', FALSE
+                        SELECT $1, $2, unnest($3::timestamptz[]), unnest($4::numeric[]), unnest($5::numeric[]), unnest($6::numeric[]), unnest($7::numeric[]), unnest($8::bigint[]), $9, 'Yahoo', FALSE
                         ON CONFLICT (futures_setting_id, (COALESCE(contract_month, '')), time, timeframe, is_continuous) DO NOTHING
                     `;
                     const params = [futuresSettingId, contractMonth, times, opens, highs, lows, closes, volumes, timeframe];
@@ -1043,7 +1043,7 @@ async function storeHistoricalData(client, futuresSettingId, contractMonth, time
             unnest($5::numeric[]),                  -- high
             unnest($6::numeric[]),                  -- low
             unnest($7::numeric[]),                  -- close
-            unnest($8::integer[]),                  -- volume
+            unnest($8::bigint[]),                   -- volume
             $9,                                     -- timeframe
             'IBKR',                                 -- source
             FALSE                                   -- is_continuous
@@ -1721,7 +1721,7 @@ async function _buildAndStoreContinuousSeries(client, futuresSettingId, timefram
     // pattern already used for Yahoo/IBKR contract writes elsewhere in this file.
     await client.query(
         `INSERT INTO historical_data (futures_setting_id, contract_month, time, open, high, low, close, volume, timeframe, source, is_continuous, is_rollover, rollover_type)
-         SELECT $1, unnest($2::varchar[]), unnest($3::timestamptz[]), unnest($4::numeric[]), unnest($5::numeric[]), unnest($6::numeric[]), unnest($7::numeric[]), unnest($8::integer[]), $9, unnest($10::varchar[]), TRUE, unnest($11::boolean[]), unnest($12::varchar[])
+         SELECT $1, unnest($2::varchar[]), unnest($3::timestamptz[]), unnest($4::numeric[]), unnest($5::numeric[]), unnest($6::numeric[]), unnest($7::numeric[]), unnest($8::bigint[]), $9, unnest($10::varchar[]), TRUE, unnest($11::boolean[]), unnest($12::varchar[])
          ON CONFLICT (futures_setting_id, (COALESCE(contract_month, '')), time, timeframe, is_continuous)
          DO UPDATE SET
             open = EXCLUDED.open, high = EXCLUDED.high, low = EXCLUDED.low, close = EXCLUDED.close,
